search_server/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
accelerate==0.34.0
sentence-transformers==3.0.1
boto3==1.35.11
requests-aws4auth==1.3.1
opensearch-haystack==0.9.0
